# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CyberArcade CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /segments:
    get:
      operationId: getSegments
      tags: [segments]
      summary: List segments
      responses:
        "200":
          description: Segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Segment"
    post:
      operationId: createSegment
      tags: [segments]
      summary: Create a segment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSegmentInput"
      responses:
        "201":
          description: Created segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
  /segments/{id}:
    get:
      operationId: getSegment
      tags: [segments]
      summary: Get a segment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
    patch:
      operationId: updateSegment
      tags: [segments]
      summary: Update a segment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSegmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
    delete:
      operationId: deleteSegment
      tags: [segments]
      summary: Delete a segment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /segments/{id}/preview:
    post:
      operationId: previewSegment
      tags: [segments]
      summary: Preview segment contact count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SegmentFilters"
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentPreviewResult"
  /segments/preview:
    post:
      operationId: previewNewSegment
      tags: [segments]
      summary: Preview a new segment contact count
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SegmentFilters"
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentPreviewResult"
...
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Segment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        filters:
          type: object
        contactCount:
          type: integer
        color:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - filters
        - contactCount
        - color
        - createdAt
    SegmentFilters:
      type: object
      properties:
        match:
          type: string
          enum: [all, any]
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/SegmentCondition"
      required:
        - match
        - conditions
    SegmentCondition:
      type: object
      properties:
        field:
          type: string
        operator:
          type: string
        value:
          type: ["string", "null"]
      required:
        - field
        - operator
    SegmentPreviewResult:
      type: object
      properties:
        count:
          type: integer
      required:
        - count
    CreateSegmentInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        filters:
          type: object
        color:
          type: string
      required:
        - name
        - filters
        - color
    UpdateSegmentInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        filters:
          type: object
        color:
          type: string
...