# lib/api-zod/src/generated/types/segment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberArcade CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { SegmentFiltersProperty } from "./segmentFiltersProperty";

export interface Segment {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  filters: SegmentFiltersProperty;
  contactCount: number;
  color: string;
  createdAt: string;
}