# lib/api-zod/src/generated/types/updateSegmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberArcade CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSegmentInputFilters } from "./updateSegmentInputFilters";

export interface UpdateSegmentInput {
  name?: string;
  /** @nullable */
  description?: string | null;
  filters?: UpdateSegmentInputFilters;
  color?: string;
}